// core/more-details/interface/card-more-details.interface.ts
export interface CardMoreDetails {
    id: string;
    titulo: string;
    prioridad: string;
    estado: string;
    creador: string;
    solucionador: string;
    fechaSolicitud: string;
    fechaTerminacion: string;
    jefeDirecto: string;
    usuariosAsignados: string;
    unidad: string;
    area: string;
    departamento: string;
    acabados: string;
    electricidad: string;
    nombreCuestionario: string;
    pregunta: string;
}